Web/clicker/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
gunicorn==21.2.0
